requirements.txt
# Used to install pinned dependencies
# Useful for dev/test jobs caches
# Must be kept in sync with setup.py
torch == 2.2.0 # This is what is installed in CI today
ibm-fms >= 0.0.3
transformers >= 4.35.0
accelerate >= 0.26.1
